+++ libs/py/tf/tfvars.py
@@ -72,6 +72,7 @@ class Cloud(BaseTfVarsModel):
 class Env(BaseTfVarsModel):
     name: str
     short_name: str
+    initial_start: bool = False
     users: dict[str, User]
     apps: dict[str, App]
     import_secrets: dict[str, ImportSecret]